+++ package.json
@@ -4,14 +4,31 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "lint": "eslint .",
     "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "lint-staged": {
+    "*.js": [
+      "npm run lint -- --fix",
+      "git add"
+    ]
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "axios": "^1.4.0",
     "cors": "^2.8.5",
+    "eslint": "^8.44.0",
     "express": "^4.18.2",
     "sqlite3": "^5.1.6"
+  },
+  "devDependencies": {
+    "lint-staged": "^13.2.3"
   }
 }
